requirements.txt
# AI/LLM Libraries
openai>=1.0.0
anthropic>=0.18.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.0.0

# Optional: for advanced features
colorama>=0.4.6  # For colored console output
